pyproject.toml
[tool.poetry]
name = "back-for-hyun-shop"
version = "0.1.0"
description = ""
authors = ["hyunsokstar <hyunsokstar@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.1.7"
django-debug-toolbar = "^4.2.0"
django-environ = "^0.11.2"
django-bootstrap5 = "^23.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
